+++ experiments_ovary_centres/gui_annot_center_correction.py
@@ -45,8 +45,8 @@ from matplotlib.figure import Figure
 from matplotlib.backends.backend_gtkagg import FigureCanvasGTKAgg
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')] # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.drawing as tl_visu
 
 PATH_BASE = tl_data.update_path(os.path.join('data_images', 'drosophila_ovary_slice'))
 PATH_IMAGES = os.path.join(PATH_BASE, 'image', '*.jpg')
