+++ experiments_ovary_centres/run_center_clustering.py
@@ -26,9 +26,9 @@ if os.environ.get('DISPLAY', '') == '' \
 import matplotlib.pylab as plt
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.drawing as tl_visu
 import run_center_candidate_training as run_train
 # import run_center_prediction as run_pred
