+++ experiments_ovary_centres/run_center_prediction.py
@@ -23,8 +23,8 @@ from functools import partial
 import pandas as pd
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')] # Add path to root
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.data_io as tl_data
 import imsegm.classification as seg_clf
 import run_center_candidate_training as run_train
 import run_center_clustering as run_clust
