+++ experiments_ovary_centres/run_create_annotation.py
@@ -33,8 +33,8 @@ from skimage import morphology, measure, draw
 
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
 import run_center_candidate_training as run_train
 
 NAME_DIR = 'annot_centres'
