+++ experiments_ovary_detect/run_ellipse_annot_match.py
@@ -23,9 +23,9 @@ import pandas as pd
 import numpy as np
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.drawing as tl_visu
 # import segmentation.annotation as seg_annot
 import imsegm.ellipse_fitting as ell_fit
