+++ experiments_ovary_detect/run_ellipse_cut_scale.py
@@ -22,8 +22,8 @@ import numpy as np
 from skimage import transform
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
 import imsegm.ellipse_fitting as ell_fit
 import run_ellipse_annot_match as r_match
