+++ experiments_ovary_detect/run_export_user-annot-segm.py
@@ -34,9 +34,9 @@ import pandas as pd
 import matplotlib.pylab as plt
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.drawing as tl_visu
 import imsegm.annotation as seg_annot
 
 NB_THREADS = max(1, int(mproc.cpu_count() * 0.8))
