+++ experiments_ovary_detect/run_ovary_egg-segmentation.py
@@ -52,9 +52,9 @@ from skimage import measure, draw
 from skimage.measure.fit import EllipseModel
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.drawing as tl_visu
 import imsegm.superpixels as seg_spx
 import imsegm.region_growing as seg_rg
 import imsegm.ellipse_fitting as ell_fit
