+++ experiments_segmentation/run_eval_superpixels.py
@@ -31,9 +31,9 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
+import imsegm.utilities.drawing as tl_visu
 import imsegm.superpixels as seg_spx
 import imsegm.labeling as seg_lbs
 from run_segm_slic_model_graphcut import load_image
