+++ handling_annotations/run_image_convert_label_color.py
@@ -23,8 +23,8 @@ from functools import partial
 import numpy as np
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
 import imsegm.annotation as seg_annot
 
 PATH_INPUT = os.path.join('data_images', 'drosophila_ovary_slice', 'segm', '*.png')
