+++ handling_annotations/run_segm_annot_relabel.py
@@ -22,8 +22,8 @@ from functools import partial
 import numpy as np
 
 sys.path += [os.path.abspath('.'), os.path.abspath('..')]  # Add path to root
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.experiments as tl_expt
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.experiments as tl_expt
 
 PATH_IMAGES = os.path.join('data_images', 'drosophila_ovary_slice', 'center_levels', '*.png')
 PATH_OUTPUT = os.path.join('results', 'relabel_center_levels')
