+++ imsegm/classification.py
@@ -29,7 +29,7 @@ from sklearn import pipeline, linear_model, neural_network
 from sklearn import model_selection
 
 import imsegm.labeling as seg_lbs
-import imsegm.utils.experiments as tl_expt
+import imsegm.utilities.experiments as tl_expt
 
 # NAME_FILE_RESULTS = 'results.csv'
 TEMPLATE_NAME_CLF = 'classifier_{}.pkl'
