+++ imsegm/ellipse_fitting.py
@@ -10,7 +10,7 @@ from skimage import morphology
 
 from skimage.measure import fit as sk_fit
 # from skimage.measure.fit import EllipseModel  # fix in future skimage>0.13.0
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.drawing as tl_visu
 import imsegm.descriptors as seg_fts
 import imsegm.superpixels as seg_spx
