+++ imsegm/tests/test_descriptors.py
@@ -15,9 +15,9 @@ from skimage import draw, transform
 import matplotlib.pyplot as plt
 
 sys.path.append(os.path.abspath(os.path.join('..', '..')))  # Add path to root
-import imsegm.utils.data_samples as d_spl
-import imsegm.utils.data_io as tl_data
-import imsegm.utils.drawing as tl_visu
+import imsegm.utilities.data_samples as d_spl
+import imsegm.utilities.data_io as tl_data
+import imsegm.utilities.drawing as tl_visu
 import imsegm.descriptors as seg_fts
 import imsegm.superpixels as seg_spx
