+++ imsegm/tests/test_graph-cut.py
@@ -13,8 +13,8 @@ import numpy as np
 import pandas as pd
 
 sys.path.append(os.path.abspath(os.path.join('..', '..')))  # Add path to root
-import imsegm.utils.data_samples as d_spl
-import imsegm.utils.data_io as tl_data
+import imsegm.utilities.data_samples as d_spl
+import imsegm.utilities.data_io as tl_data
 import imsegm.superpixels as seg_spx
 import imsegm.graph_cuts as seg_gc
 import imsegm.labeling as seg_lb
