+++ imsegm/utilities/drawing.py
