+++ imsegm/utilities/experiments.py
